feat(raw-client): separate raw-client from substrate-client
