+++ packages/raw-client/README.md
@@ -1,0 +1,3 @@
+# @polkadot-api/raw-client
+
+Very low-level stuff, use it at your own risk...
