+++ packages/raw-client/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./RpcError"
 export * from "./createClient"
 export * from "./DestroyedError"
+export * from "./subscriptions-manager"
