+++ packages/substrate-client/src/index.ts
@@ -2,13 +2,13 @@ export { AbortError } from "@polkadot-api/utils"
 export type * from "@polkadot-api/json-rpc-provider"
 
 export type * from "./common-types"
-export type * from "./client"
+export type * from "@polkadot-api/raw-client"
 export type * from "./transaction"
 export type * from "./chainhead"
 export type * from "./chainspec"
 export type * from "./archive"
 
-export { RpcError, DestroyedError } from "./client"
+export { RpcError, DestroyedError } from "@polkadot-api/raw-client"
 export {
   StopError,
   DisjointError,
