+++ packages/substrate-client/src/internal-utils/index.ts
@@ -1,4 +1,3 @@
 export * from "./abortablePromiseFn"
 export * from "./deferred-promise"
 export * from "./noop"
-export * from "./subscriptions-manager"
